ng: add scripts for msm core_ctl

+++ ramdisk-ng/sbin/init.cocore.post_boot.sh
@@ -96,6 +96,21 @@ echo "90 1747200:80" > /sys/devices/system/cpu/cpu4/cpufreq/interactive/target_l
 echo 10000 > /sys/devices/system/cpu/cpu4/cpufreq/interactive/timer_rate
 echo 40000 > /sys/devices/system/cpu/cpu4/cpufreq/interactive/timer_slack
 
+# MSM Core Control Settings
+if [ -e /system/lib/modules/core_ctl.ko ]; then
+  insmod /system/lib/modules/core_ctl.ko
+
+  # Power cluster
+
+  # Perf cluster
+  echo 2 >   /sys/devices/system/cpu/cpu4/core_ctl/min_cpus
+  echo 4 >   /sys/devices/system/cpu/cpu4/core_ctl/max_cpus
+  echo 68 >  /sys/devices/system/cpu/cpu4/core_ctl/busy_up_thres
+  echo 40 >  /sys/devices/system/cpu/cpu4/core_ctl/busy_down_thres
+  echo 100 > /sys/devices/system/cpu/cpu4/core_ctl/offline_delay_ms
+  echo 1 >   /sys/devices/system/cpu/cpu4/core_ctl/is_big_cluster
+fi
+
 # Re-Enable thermal & BCL core_control now
 echo 1 > /sys/module/msm_thermal/core_control/enabled
 for mode in /sys/devices/soc.0/qcom,bcl.*/mode
@@ -122,8 +137,10 @@ done
 start perfd
 
 # CPUQuiet Governor Settings
-echo 4 > /sys/devices/system/cpu/cpuquiet/nr_min_cpus
-echo rqbalance > /sys/devices/system/cpu/cpuquiet/current_governor
+if [ -e /sys/devices/system/cpu/cpuquiet ]; then
+  echo 4 > /sys/devices/system/cpu/cpuquiet/nr_min_cpus
+  echo rqbalance > /sys/devices/system/cpu/cpuquiet/current_governor
+fi
 
 #
 # CPUFreq Settings: end
